pyproject.toml
[tool.poetry]
name = "ml_engineering_lab4"
description = "Training pipeline for deep learning model"
authors = ["amex.ua@hotmail.com"]
version = "0.01"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
dvc = "^3.59"
torch = "^2.6.0"
torchvision = "^0.21.0"
scikit-learn = "^1.6"
tqdm = "^4.67"
matplotlib = "^3.10"
numpy = "^2.2"
pyyaml = "^6.0"
scipy = "^1.15"
pandas = "^2.2"
requests = "^2.0"
mlflow = "^2.22"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15"
ruff = "^0.11"
black = "^25.0"
isort = "^6.0"
ipykernel = "^6.29"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
